module2-task1 - Валидируй, как никогда не валидировал  - Создан компонент place-card.jsx  - Указаны параметры пропсов, применяемых для компонентов place-card.jsx, app.jsx, main.jsx;  - Данные по доступным предложениям передаются в приложение из index.js
